+++ Payload_Type/apollo/mythic/agent_functions/inline_assembly.py
@@ -48,9 +48,9 @@ class InlineAssemblyCommand(CommandBase):
                                               delete_after_fetch=True)
         if file_resp.status == MythicStatus.Success:
             task.args.add_arg("loader_stub_id", file_resp.response['agent_file_id'])
+            task.display_params = "Running inline_assembly: {}".format(path.basename(dllPath))
         else:
             raise Exception("Failed to register execute-assembly DLL: " + file_resp.error)
-
         return task
 
     async def process_response(self, response: AgentResponse):
